feat: implement validations in services

+++ src/LionHeart.BusinessLogic/Services/BasketEntryService.cs
@@ -1,30 +1,47 @@
 ﻿using LionHeart.BusinessLogic.Resources;
+using LionHeart.BusinessLogic.FluentValidations.Validators.BasketEntry;
+using LionHeart.BusinessLogic.FluentValidations.Models;
 using LionHeart.Core.Dtos.BasketEntry;
 using LionHeart.Core.Interfaces.Repositories;
 using LionHeart.Core.Interfaces.Services;
 using LionHeart.Core.Models;
-using LionHeart.Core.Result;
+using LionHeart.Core.Results;
+using LionHeart.Core.ValidationModels.BasketEntry;
 
 namespace LionHeart.BusinessLogic.Services;
 
 public class BasketEntryService : IBasketEntryService
 {
     private readonly IBasketEntryRepository _basketEntryRepository;
+    private readonly IProductRepository _productRepository;
+    private readonly BasketEntryServiceValidators _validators;
 
-    public BasketEntryService(IBasketEntryRepository basketEntryRepository)
+    public BasketEntryService(IBasketEntryRepository basketEntryRepository,
+                              IProductRepository productRepository,
+                              BasketEntryServiceValidators validators)
     {
         _basketEntryRepository = basketEntryRepository;
+        _productRepository = productRepository;
+        _validators = validators;
     }
 
     public async Task<Result<BasketEntry>> GetById(string id)
     {
         try
         {
+            var idValidationResult = _validators.IdValidator.Validate(new IdModel(id));
+            if (!idValidationResult.IsValid)
+            {
+                var errorMessages = idValidationResult.Errors.Select(e => e.ErrorMessage);
+                return Result<BasketEntry>.Failure(errorMessages);
+            }
+
             var entry = await _basketEntryRepository.GetById(id);
             if (entry is null)
             {
                 return Result<BasketEntry>.Failure(ErrorMessage.BasketEntryNotFound);
             }
+
             return Result<BasketEntry>.Success(entry);
         }
         catch
@@ -36,6 +53,13 @@ public class BasketEntryService : IBasketEntryService
     {
         try
         {
+            var idValidationResult = _validators.IdValidator.Validate(new IdModel(userId, productId));
+            if (!idValidationResult.IsValid)
+            {
+                var errorMessages = idValidationResult.Errors.Select(e => e.ErrorMessage);
+                return Result<BasketEntry>.Failure(errorMessages);
+            }
+
             var entry = await _basketEntryRepository.GetByAlternateKey(userId, productId);
             if (entry is null)
             {
@@ -52,6 +76,13 @@ public class BasketEntryService : IBasketEntryService
     {
         try
         {
+            var idValidationResult = _validators.IdValidator.Validate(new IdModel(userId));
+            if (!idValidationResult.IsValid)
+            {
+                var errorMessages = idValidationResult.Errors.Select(e => e.ErrorMessage);
+                return Result<List<BasketEntry>>.Failure(errorMessages);
+            }
+
             var entries = await _basketEntryRepository.GetEntriesByUserId(userId);
             return Result<List<BasketEntry>>.Success(entries);
         }
@@ -64,17 +95,39 @@ public class BasketEntryService : IBasketEntryService
     {
         try
         {
-            var entry = new BasketEntry
+            var dtoValidationResult = _validators.AddBasketEntryDtoValidator.Validate(dto);
+            if (!dtoValidationResult.IsValid)
+            {
+                var errorMessages = dtoValidationResult.Errors.Select(e => e.ErrorMessage);
+                return Result<BasketEntry>.Failure(errorMessages);
+            }
+
+            bool basketEntryAlreadyExists = await _basketEntryRepository.Exists(dto.UserId, dto.ProductId);
+            bool productExists = await _productRepository.Exists(p => p.Id == dto.ProductId);
+
+            var validateAddModel = new ValidateAddModel()
+            {
+                BasketEntryAlreadyExists = basketEntryAlreadyExists,
+                ProductExists = productExists
+            };
+            var basketEntryValidatorResult = _validators.BasketEntryValidator.ValidateAdd(validateAddModel);
+            if (basketEntryValidatorResult.IsFaulted)
+            {
+                return Result<BasketEntry>.Failure(basketEntryValidatorResult.ErrorMessages);
+            }
+
+            var entry = new BasketEntry()
             {
                 UserId = dto.UserId,
                 ProductId = dto.ProductId,
                 CreatedAt = dto.CreatedAt
             };
-            bool entryAlreadyExists = await _basketEntryRepository.Exists(dto.UserId, dto.ProductId);
-            if (entryAlreadyExists) return Result<BasketEntry>.Failure(ErrorMessage.BasketEntryAlreadyExists);
-            var result = await _basketEntryRepository.Add(entry);
-            if (result <= 0) return Result<BasketEntry>.Failure(ErrorMessage.BasketEntryNotCreated);
-            
+            var basketEntryRepositoryResult = await _basketEntryRepository.Add(entry);
+            if (basketEntryRepositoryResult <= 0)
+            {
+                return Result<BasketEntry>.Failure(ErrorMessage.BasketEntryNotCreated);
+            }
+
             return Result<BasketEntry>.Success(entry);
         }
         catch
@@ -86,14 +139,32 @@ public class BasketEntryService : IBasketEntryService
     {
         try
         {
+            var dtoValidationResult = _validators.UpdateBasketEntryDtoValidator.Validate(dto);
+            if (!dtoValidationResult.IsValid)
+            {
+                var errorMessages = dtoValidationResult.Errors.Select(e => e.ErrorMessage);
+                return Result<BasketEntry>.Failure(errorMessages);
+            }
+
+            bool basketEntryExists = await _basketEntryRepository.Exists(e => e.Id == dto.Id);
+            var validateUpdateModel = new ValidateUpdateModel()
+            {
+                BasketEntryExists = basketEntryExists,
+            };
+            var basketEntryValidatorResult = _validators.BasketEntryValidator.ValidateUpdate(validateUpdateModel);
+            if (basketEntryValidatorResult.IsFaulted)
+            {
+                return Result<BasketEntry>.Failure(basketEntryValidatorResult.ErrorMessages);
+            }
+
             var entry = await _basketEntryRepository.GetById(dto.Id);
             if (entry is null)
             {
                 return Result<BasketEntry>.Failure(ErrorMessage.BasketEntryNotFound);
             }
             entry.Quantity = dto.Quantity;
-            var result = await _basketEntryRepository.Update(entry);
-            if (result <= 0)
+            var basketEntryRepositoryResult = await _basketEntryRepository.Update(entry);
+            if (basketEntryRepositoryResult <= 0)
             {
                 return Result<BasketEntry>.Failure(ErrorMessage.BasketEntryNotUpdated);
             }
@@ -108,13 +179,31 @@ public class BasketEntryService : IBasketEntryService
     {
         try
         {
+            var idValidationResult = _validators.IdValidator.Validate(new IdModel(id));
+            if (!idValidationResult.IsValid)
+            {
+                var errorMessages = idValidationResult.Errors.Select(e => e.ErrorMessage);
+                return Result<BasketEntry>.Failure(errorMessages);
+            }
+
+            bool basketEntryExists = await _basketEntryRepository.Exists(e => e.Id == id);
+            var validateRemoveModel = new ValidateRemoveModel()
+            {
+                BasketEntryExists = basketEntryExists,
+            };
+            var basketEntryValidatorResult = _validators.BasketEntryValidator.ValidateRemove(validateRemoveModel);
+            if (basketEntryValidatorResult.IsFaulted)
+            {
+                return Result<BasketEntry>.Failure(basketEntryValidatorResult.ErrorMessages);
+            }
+
             var entry = await _basketEntryRepository.GetById(id);
             if (entry is null)
             {
                 return Result<BasketEntry>.Failure(ErrorMessage.BasketEntryNotFound);
             }
-            var result = await _basketEntryRepository.Remove(entry);
-            if (result <= 0)
+            var basketEntryRepositoryResult = await _basketEntryRepository.Remove(entry);
+            if (basketEntryRepositoryResult <= 0)
             {
                 return Result<BasketEntry>.Failure(ErrorMessage.BasketEntryNotDeleted);
             }
@@ -129,13 +218,20 @@ public class BasketEntryService : IBasketEntryService
     {
         try
         {
+            var idValidationResult = _validators.IdValidator.Validate(new IdModel(ids));
+            if (!idValidationResult.IsValid)
+            {
+                var errorMessages = idValidationResult.Errors.Select(e => e.ErrorMessage);
+                return Result<List<BasketEntry>>.Failure(errorMessages);
+            }
+
             var entries = await _basketEntryRepository.Find(ids);
-            if (entries.Count == 0)
+            if (entries.Count == 0 || entries.Count != ids.Count)
             {
                 return Result<List<BasketEntry>>.Failure(ErrorMessage.BasketEntriesNotFound);
             }
-            var result = await _basketEntryRepository.RemoveRange(entries);
-            if (result <= 0)
+            var basketEntryRepositoryResult = await _basketEntryRepository.RemoveRange(entries);
+            if (basketEntryRepositoryResult <= 0)
             {
                 return Result<List<BasketEntry>>.Failure(ErrorMessage.BasketEntriesNotRemoved);
             }
@@ -150,8 +246,15 @@ public class BasketEntryService : IBasketEntryService
     {
         try
         {
-            var result = await _basketEntryRepository.Exists(userId, productId);
-            return Result<bool>.Success(result);
+            var idValidationResult = _validators.IdValidator.Validate(new IdModel(userId, productId));
+            if (!idValidationResult.IsValid)
+            {
+                var errorMessages = idValidationResult.Errors.Select(e => e.ErrorMessage);
+                return Result<bool>.Failure(errorMessages);
+            }
+
+            bool basketEntryExists = await _basketEntryRepository.Exists(userId, productId);
+            return Result<bool>.Success(basketEntryExists);
         }
         catch
         {
